% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION1.R
\name{FUNCTION}
\alias{FUNCTION}
\title{Retrieve GBIF occurrence data and push it to Geospatial Quality API}
\usage{
FUNCTION()
}
\value{
Returns the data retrieved in string format and the pushed data to Geospatial Quality API.
}
\description{
\code{FUNCTION} Function to retrieve GBIF occurrence data of all terrestrial mammals in Australia and then sends all of them to the Geospatial Quality API.
}
\examples{
FUNCTION()
}
